tex creation + renaming + moar gl calls
