+++ src/GLEngine/GLTexture.cpp
@@ -13,10 +13,10 @@ CGLTexture::~CGLTexture()
 void CGLTexture::Draw(int x, int y, bool checktrans)
 {
     DEBUG_TRACE_FUNCTION;
+#ifndef NEW_RENDERER_ENABLED
     if (Texture == 0)
         return;
 
-#ifndef NEW_RENDERER_ENABLED
     if (checktrans)
     {
         glEnable(GL_BLEND);
@@ -32,15 +32,17 @@ void CGLTexture::Draw(int x, int y, bool checktrans)
         g_GL_Draw(*this, x, y);
     }
 #else
-    auto cmd = DrawQuadCmd(Texture, x, y, Width, Height, 1.f, 1.f);
+    if (Texture == RENDER_TEXTUREHANDLE_INVALID)
+        return;
 
+    auto cmd = DrawQuadCmd(Texture, x, y, Width, Height, 1.f, 1.f);
     if (checktrans)
     {
         RenderAdd_SetBlend(g_renderCmdList, &BlendStateCmd(BlendFunc::SrcAlpha_OneMinusSrcAlpha));
         RenderAdd_DrawQuad(g_renderCmdList, &cmd, 1);
         RenderAdd_DisableBlend(g_renderCmdList);
 
-        // FIXME what were the original func, op, and stencil values?
+        // FIXME what are the assumed values for func, op, ref, and mask?
         RenderAdd_SetStencil(g_renderCmdList, &StencilStateCmd(StencilFunc::Greater));
         RenderAdd_DrawQuad(g_renderCmdList, &cmd, 1);
         RenderAdd_DisableStencil(g_renderCmdList);
@@ -55,8 +57,14 @@ void CGLTexture::Draw(int x, int y, bool checktrans)
 void CGLTexture::Draw(int x, int y, int width, int height, bool checktrans)
 {
     DEBUG_TRACE_FUNCTION;
+#ifndef NEW_RENDERER_ENABLED
     if (Texture == 0)
+#else
+    if (Texture == RENDER_TEXTUREHANDLE_INVALID)
+#endif
+    {
         return;
+    }
 
     if (width == 0)
     {
@@ -119,14 +127,19 @@ void CGLTexture::Draw_Tooltip(int x, int y, int width, int height)
 void CGLTexture::DrawRotated(int x, int y, float angle)
 {
     DEBUG_TRACE_FUNCTION;
+#ifndef NEW_RENDERER_ENABLED
     if (Texture == 0)
     {
         return;
     }
 
-#ifndef NEW_RENDERER_ENABLED
     g_GL_DrawRotated(*this, x, y, angle);
 #else
+    if (Texture == RENDER_TEXTUREHANDLE_INVALID)
+    {
+        return;
+    }
+
     auto cmd = DrawRotatedQuadCmd(Texture, x, y - Height, Width, Height, angle);
     RenderAdd_DrawRotatedQuad(g_renderCmdList, &cmd, 1);
 #endif
@@ -135,12 +148,12 @@ void CGLTexture::DrawRotated(int x, int y, float angle)
 void CGLTexture::DrawTransparent(int x, int y, bool stencil)
 {
     DEBUG_TRACE_FUNCTION;
+#ifndef NEW_RENDERER_ENABLED
     if (Texture == 0)
     {
         return;
     }
 
-#ifndef NEW_RENDERER_ENABLED
     glEnable(GL_BLEND);
     glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
     glColor4f(1.0f, 1.0f, 1.0f, 0.25f);
@@ -154,6 +167,11 @@ void CGLTexture::DrawTransparent(int x, int y, bool stencil)
         glDisable(GL_STENCIL_TEST);
     }
 #else
+    if (Texture == RENDER_TEXTUREHANDLE_INVALID)
+    {
+        return;
+    }
+
     auto cmd = DrawQuadCmd(Texture, x, y, Width, Height, 1.f, 1.f, { { 1.f, 1.f, 1.f, 0.25f } });
     RenderAdd_SetBlend(g_renderCmdList, &BlendStateCmd(BlendFunc::SrcAlpha_OneMinusSrcAlpha));
     RenderAdd_DrawQuad(g_renderCmdList, &cmd, 1);
@@ -172,11 +190,19 @@ void CGLTexture::DrawTransparent(int x, int y, bool stencil)
 void CGLTexture::Clear()
 {
     DEBUG_TRACE_FUNCTION;
+#ifndef NEW_RENDERER_ENABLED
     if (Texture != 0)
     {
         glDeleteTextures(1, &Texture);
         Texture = 0;
     }
+#else
+    if (Texture != RENDER_TEXTUREHANDLE_INVALID)
+    {
+        Render_DestroyTexture(Texture);
+        Texture = RENDER_TEXTUREHANDLE_INVALID;
+    }
+#endif
 
     if (VertexBuffer != 0)
     {
