+++ src/GLEngine/GLTexture.h
@@ -4,6 +4,8 @@
 #pragma once
 
 #include "GLHeaders.h"
+// FIXME only RenderTypes.h should be needed here, RenderAPI is included due to temp NEW_RENDERER_ENABLED
+#include "Renderer/RenderAPI.h"
 
 struct CGLTexture
 {
@@ -11,7 +13,11 @@ struct CGLTexture
     uint16_t Height = 0;
     GLuint VertexBuffer = 0;
     GLuint MirroredVertexBuffer = 0;
+#ifndef NEW_RENDERER_ENABLED
     GLuint Texture = 0;
+#else
+    texture_handle_t Texture = RENDER_TEXTUREHANDLE_INVALID;
+#endif
 
     void Draw_Tooltip(int x, int y, int width, int height);
     virtual void Draw(int x, int y, bool checktrans = false);
