+++ src/Renderer/RenderInternal.h
@@ -1,3 +1,6 @@
+// GPLv3 License
+// Copyright (C) 2019 Everton Fernando Patitucci da Silva
+
 #pragma once
 
 #ifndef RENDERER_INTERNAL
